[guppy] move core select logic into a select_core module
This logic is shared between package and feature graphs.
